spine model updated
