+++ Sources/Spine/Model/Animation/Keyframes/Slot/SlotKeyframeColorModel.swift
@@ -11,7 +11,7 @@ struct SlotKeyframeColorModel: CurvedKeyframeModel {
     
     let time: TimeInterval
     let color: ColorModel
-    var curve: CurveModelType
+    var curve: CurveModel
 }
 
 extension SlotKeyframeColorModel: Decodable {
@@ -26,6 +26,6 @@ extension SlotKeyframeColorModel: Decodable {
         let container = try decoder.container(keyedBy: Keys.self)
         time = try container.decodeIfPresent(TimeInterval.self, forKey: .time) ?? 0
         color = try container.decode(ColorModel.self, forKey: .color)
-        curve = try container.decodeIfPresent(CurveModelType.self, forKey: .curve) ?? .linear
+        curve = try container.decodeIfPresent(CurveModel.self, forKey: .curve) ?? .linear
     }
 }
