+++ Sources/Spine/Model/Protocols/SpineDecodableArray.swift
@@ -1,0 +1,13 @@
+//
+//  File.swift
+//  
+//
+//  Created by Max Gribov on 12.11.2022.
+//
+
+import Foundation
+
+protocol SpineDecodableArray {
+    
+    init(_ name: String, _ container: inout UnkeyedDecodingContainer) throws
+}
