+++ Sources/Spine/Model/Types/BezierCurveModel.swift
@@ -1,0 +1,16 @@
+//
+//  BezierCurveModel.swift
+//  
+//
+//  Created by Max Gribov on 12.11.2022.
+//
+
+import Foundation
+
+struct BezierCurveModel: Equatable {
+    
+    let p0: Float
+    let p1: Float
+    let p2: Float
+    let p3: Float
+}
