+++ Sources/Spine/Utilities/Constants.swift
@@ -1,0 +1,10 @@
+//
+//  Constants.swift
+//  
+//
+//  Created by Max Gribov on 12.11.2022.
+//
+
+import SpriteKit
+
+let degreeToRadiansFactor: CGFloat = CGFloat.pi / 180.0
